+++ docs/CHANGELOG.md
@@ -43,6 +43,7 @@ Semantic Versioning.
 * TypeScript support (still experimental):
     * `let x: C<T>=y;` no longer falsely reports an error.
     * `class A extends B<C> {}` no longer falsely reports an error.
+    * `T extends keyof O ? A : B` no longer falsely reports an error.
 
 ### Changed